Remove artifactName parameter from download artifacts task
When specific artifact name is specified for DownloadPipelineArtifact, then files are downloaded to the root of $(Pipeline.Workspace) directory.
When specific artifact name is not specified, then files are downloaded to specific artifacts folders, e.g. $(Pipeline.Workspace)/some-artifact.

We use the 2nd approach which avoids collisions if multiple artifacts are used.